requirements.txt
pandas==1.5.3  
numpy==1.24.2  
apache-airflow==2.6.3  
apache-airflow-providers-postgres==5.4.0  
elasticsearch==8.5.3  
